function/requirements.txt
azure-functions
psycopg2-binary
sendgrid==6.9.7